fix: armazena preço do produto em centavos e converte na serialização

+++ src/main/java/com/projedata/myindustry/dto/ProductRequest.java
@@ -1,3 +1,3 @@
 package com.projedata.myindustry.dto;
 
-public record ProductRequest(String name, Integer price) {}
+public record ProductRequest(String name, Double price) {}
