+++ src/main/java/com/projedata/myindustry/dto/ProductionSuggestionResponse.java
@@ -2,4 +2,4 @@ package com.projedata.myindustry.dto;
 
 import java.util.List;
 
-public record ProductionSuggestionResponse(List<ProductionItemResponse> items, Integer grandTotal) {}
+public record ProductionSuggestionResponse(List<ProductionItemResponse> items, Double grandTotal) {}
